move workflow stopContainer logic to base class
